+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 1.0.0 (11/09/2017)
+## 2.0.0 (4/10/2018)
+- Created new modules to support both JUnit4 and JUnit 5.
+
+## 1.0.0 (09/11/2017)
 - Initial release!
 - Based off of Kafka Server and Kafka-Clients version 0.11.0.1
+- Built for JUnit 4.x